Add video transition and title messages.
One user said it was unclear whether or not a video was mirrored or not, so now there is a title message and also the videos that swap animate. making it really obvious whether or not the videos are mirrored.

TODO: Add an option to disable the animation, because some users might not like it.